requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5
backcall==0.2.0
boto3==1.17.102
botocore==1.20.102
colorama==0.4.4
decorator==5.0.9
distlib==0.3.2
Django==3.1.7
django-adminlte3==0.1.6
django-crispy-forms==1.11.2
django-filter==2.4.0
django-formtools==2.3
django-passwords==0.3.12
django-phonenumber-field==5.0.0
django-storages==1.11.1
djangorestframework==3.12.2
dynamic-django-forms==0.1.5
filelock==3.0.12
graphviz==0.16
gunicorn==20.1.0
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
parso==0.8.2
phonenumbers==8.12.20
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.19
Pygments==2.9.0
pylint==2.7.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.6
virtualenv==20.4.7
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
